waste_classifier_main/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
tensorflow==2.13.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
python-socketio==5.9.0
gunicorn